Move chunk code into its own package, and ChunkSection into its own file
