+++ src/main/java/net/glowstone/GlowWorld.java
@@ -4,13 +4,15 @@ import com.flowpowered.network.Message;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import net.glowstone.ChunkManager.ChunkLock;
-import net.glowstone.GlowChunk.ChunkSection;
-import net.glowstone.GlowChunk.Key;
-import net.glowstone.GlowChunkSnapshot.EmptySnapshot;
 import net.glowstone.block.GlowBlock;
 import net.glowstone.block.ItemTable;
 import net.glowstone.block.blocktype.BlockType;
+import net.glowstone.chunk.ChunkManager;
+import net.glowstone.chunk.ChunkSection;
+import net.glowstone.chunk.GlowChunk;
+import net.glowstone.chunk.ChunkManager.ChunkLock;
+import net.glowstone.chunk.GlowChunk.Key;
+import net.glowstone.chunk.GlowChunkSnapshot.EmptySnapshot;
 import net.glowstone.constants.*;
 import net.glowstone.entity.*;
 import net.glowstone.entity.objects.GlowFallingBlock;
