+++ src/main/java/net/glowstone/block/blocktype/BlockDropper.java
@@ -1,9 +1,9 @@
 package net.glowstone.block.blocktype;
 
-import net.glowstone.GlowChunk;
 import net.glowstone.block.GlowBlock;
 import net.glowstone.block.entity.TEDropper;
 import net.glowstone.block.entity.TileEntity;
+import net.glowstone.chunk.GlowChunk;
 import net.glowstone.inventory.MaterialMatcher;
 import net.glowstone.inventory.ToolType;
