+++ src/main/java/net/glowstone/block/blocktype/BlockFlowerPot.java
@@ -1,11 +1,11 @@
 package net.glowstone.block.blocktype;
 
-import net.glowstone.GlowChunk;
 import net.glowstone.block.GlowBlock;
 import net.glowstone.block.GlowBlockState;
 import net.glowstone.block.entity.TEFlowerPot;
 import net.glowstone.block.entity.TileEntity;
 import net.glowstone.block.state.GlowFlowerPot;
+import net.glowstone.chunk.GlowChunk;
 import net.glowstone.entity.GlowPlayer;
 import org.bukkit.GrassSpecies;
 import org.bukkit.Material;
