+++ src/main/java/net/glowstone/block/entity/TileEntity.java
@@ -1,8 +1,8 @@
 package net.glowstone.block.entity;
 
-import net.glowstone.GlowChunk.Key;
 import net.glowstone.block.GlowBlock;
 import net.glowstone.block.GlowBlockState;
+import net.glowstone.chunk.GlowChunk.Key;
 import net.glowstone.entity.GlowPlayer;
 import net.glowstone.util.nbt.CompoundTag;
 import org.bukkit.block.Block;
