+++ src/main/java/net/glowstone/chunk/GlowChunk.java
@@ -1,9 +1,12 @@
-package net.glowstone;
+package net.glowstone.chunk;
 
 import com.flowpowered.network.util.ByteBufUtils;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import lombok.Data;
+import net.glowstone.EventFactory;
+import net.glowstone.GlowServer;
+import net.glowstone.GlowWorld;
 import net.glowstone.block.GlowBlock;
 import net.glowstone.block.GlowBlockState;
 import net.glowstone.block.ItemTable;
@@ -11,7 +14,6 @@ import net.glowstone.block.blocktype.BlockType;
 import net.glowstone.block.entity.TileEntity;
 import net.glowstone.entity.GlowEntity;
 import net.glowstone.net.message.play.game.ChunkDataMessage;
-import net.glowstone.util.NibbleArray;
 import net.glowstone.util.VariableValueArray;
 import net.glowstone.util.nbt.CompoundTag;
 import org.bukkit.Chunk;
@@ -37,7 +39,11 @@ public final class GlowChunk implements Chunk {
     /**
      * The Y depth of a single chunk section.
      */
-    private static final int SEC_DEPTH = 16;
+    public static final int SEC_DEPTH = 16;
+    /**
+     * The number of chunk sections in a single chunk column.
+     */
+    public static final int SEC_COUNT = DEPTH / SEC_DEPTH;
     /**
      * The world of this chunk.
      */
@@ -242,7 +248,7 @@ public final class GlowChunk implements Chunk {
         }
         //GlowServer.logger.log(Level.INFO, "Initializing chunk ({0},{1})", new Object[]{x, z});
 
-        sections = new ChunkSection[DEPTH / SEC_DEPTH];
+        sections = new ChunkSection[SEC_COUNT];
         System.arraycopy(initSections, 0, sections, 0, Math.min(sections.length, initSections.length));
 
         biomes = new byte[WIDTH * HEIGHT];
@@ -624,19 +630,6 @@ public final class GlowChunk implements Chunk {
         return toMessage(skylight, true);
     }
 
-    private static final int PACKET_DATA_ARRAY_LENGTH = ChunkSection.ARRAY_SIZE * 2 / 8;
-    private static final byte[] PACKET_DATA_ARRAY_LENGTH_BYTES;
-
-    static {
-        ByteBuf buf = Unpooled.buffer();
-        try {
-            ByteBufUtils.writeVarInt(buf, PACKET_DATA_ARRAY_LENGTH);
-            PACKET_DATA_ARRAY_LENGTH_BYTES = buf.array().clone();
-        } finally {
-            buf.release();
-        }
-    }
-
     /**
      * Creates a new {@link ChunkDataMessage} which can be sent to a client to stream
      * parts of this chunk to them.
@@ -719,85 +712,6 @@ public final class GlowChunk implements Chunk {
          * The coordinates.
          */
         private final int x, z;
-    }
-
-    /**
-     * A single cubic section of a chunk, with all data.
-     */
-    public static final class ChunkSection {
-        private static final int ARRAY_SIZE = WIDTH * HEIGHT * SEC_DEPTH;
-
-        // these probably should be made non-public
-        public final char[] types;
-        public final NibbleArray skyLight;
-        public final NibbleArray blockLight;
-        public int count; // amount of non-air blocks
-
-        /**
-         * Create a new, empty ChunkSection.
-         */
-        public ChunkSection() {
-            types = new char[ARRAY_SIZE];
-            skyLight = new NibbleArray(ARRAY_SIZE);
-            blockLight = new NibbleArray(ARRAY_SIZE);
-            skyLight.fill((byte) 0xf);
-        }
-
-        /**
-         * Create a ChunkSection with the specified chunk data. This
-         * ChunkSection assumes ownership of the arrays passed in, and they
-         * should not be further modified.
-         *
-         * @param types An array of block types for this chunk section.
-         * @param skyLight An array for skylight data for this chunk section.
-         * @param blockLight An array for blocklight data for this chunk section.
-         */
-        public ChunkSection(char[] types, NibbleArray skyLight, NibbleArray blockLight) {
-            if (types.length != ARRAY_SIZE || skyLight.size() != ARRAY_SIZE || blockLight.size() != ARRAY_SIZE) {
-                throw new IllegalArgumentException("An array length was not " + ARRAY_SIZE + ": " + types.length + " " + skyLight.size() + " " + blockLight.size());
-            }
-            this.types = types;
-            this.skyLight = skyLight;
-            this.blockLight = blockLight;
-            recount();
-        }
-
-        /**
-         * Calculate the index into internal arrays for the given coordinates.
-         *
-         * @param x The x coordinate, for east and west.
-         * @param y The y coordinate, for up and down.
-         * @param z The z coordinate, for north and south.
-         *
-         * @return The index.
-         */
-        public int index(int x, int y, int z) {
-            if (x < 0 || z < 0 || x >= WIDTH || z >= HEIGHT) {
-                throw new IndexOutOfBoundsException("Coords (x=" + x + ",z=" + z + ") out of section bounds");
-            }
-            return (y & 0xf) << 8 | z << 4 | x;
-        }
-
-        /**
-         * Recount the amount of non-air blocks in the chunk section.
-         */
-        public void recount() {
-            count = 0;
-            for (char type : types) {
-                if (type != 0) {
-                    count++;
-                }
-            }
-        }
-
-        /**
-         * Take a snapshot of this section which will not reflect future changes.
-         *
-         * @return The snapshot for this section.
-         */
-        public ChunkSection snapshot() {
-            return new ChunkSection(types.clone(), skyLight.snapshot(), blockLight.snapshot());
-        }
     }
 
 }
