+++ src/main/java/net/glowstone/entity/GlowEntity.java
@@ -4,9 +4,9 @@ import com.flowpowered.network.Message;
 import com.google.common.base.Preconditions;
 import lombok.Getter;
 import net.glowstone.EventFactory;
-import net.glowstone.GlowChunk;
 import net.glowstone.GlowServer;
 import net.glowstone.GlowWorld;
+import net.glowstone.chunk.GlowChunk;
 import net.glowstone.entity.meta.MetadataIndex;
 import net.glowstone.entity.meta.MetadataIndex.StatusFlags;
 import net.glowstone.entity.meta.MetadataMap;
