+++ src/main/java/net/glowstone/entity/GlowPlayer.java
@@ -8,9 +8,7 @@ import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import lombok.Getter;
 import lombok.Setter;
-import net.glowstone.ChunkManager.ChunkLock;
 import net.glowstone.*;
-import net.glowstone.GlowChunk.Key;
 import net.glowstone.block.GlowBlock;
 import net.glowstone.block.ItemTable;
 import net.glowstone.block.blocktype.BlockBed;
@@ -18,6 +16,9 @@ import net.glowstone.block.entity.TESign;
 import net.glowstone.block.entity.TileEntity;
 import net.glowstone.block.itemtype.ItemFood;
 import net.glowstone.block.itemtype.ItemType;
+import net.glowstone.chunk.GlowChunk;
+import net.glowstone.chunk.ChunkManager.ChunkLock;
+import net.glowstone.chunk.GlowChunk.Key;
 import net.glowstone.constants.GlowAchievement;
 import net.glowstone.constants.GlowBlockEntity;
 import net.glowstone.constants.GlowEffect;
