+++ src/main/java/net/glowstone/entity/objects/GlowItemFrame.java
@@ -1,7 +1,7 @@
 package net.glowstone.entity.objects;
 
 import com.flowpowered.network.Message;
-import net.glowstone.GlowChunk.Key;
+import net.glowstone.chunk.GlowChunk.Key;
 import net.glowstone.entity.GlowEntity;
 import net.glowstone.entity.GlowPlayer;
 import net.glowstone.entity.meta.MetadataIndex;
