+++ src/main/java/net/glowstone/io/nbt/NbtStructureDataService.java
@@ -1,8 +1,8 @@
 package net.glowstone.io.nbt;
 
-import net.glowstone.GlowChunk.Key;
 import net.glowstone.GlowServer;
 import net.glowstone.GlowWorld;
+import net.glowstone.chunk.GlowChunk.Key;
 import net.glowstone.generator.structures.GlowStructure;
 import net.glowstone.io.StructureDataService;
 import net.glowstone.io.structure.StructureStorage;
