+++ src/main/java/net/glowstone/util/VariableValueArray.java
@@ -1,6 +1,6 @@
 package net.glowstone.util;
 
-public final class VariableValueArray {
+public final class VariableValueArray implements Cloneable {
 
     private final long[] backing;
     private final int capacity;
@@ -59,7 +59,7 @@ public final class VariableValueArray {
             throw new IllegalArgumentException(String.format("value (%s) must not be negative", value));
         }
         if (value > valueMask) {
-            throw new IllegalArgumentException(String.format("value (%s) must not be greater then %s", value, valueMask));
+            throw new IllegalArgumentException(String.format("value (%s) must not be greater than %s", value, valueMask));
         }
 
         index *= bitsPerValue;
@@ -80,7 +80,41 @@ public final class VariableValueArray {
             throw new IndexOutOfBoundsException(String.format("index (%s) must not be negative", index));
         }
         if (index >= capacity) {
-            throw new IndexOutOfBoundsException(String.format("index (%s) must not be greater then the capacity (%s)", index, capacity));
+            throw new IndexOutOfBoundsException(String.format("index (%s) must not be greater than the capacity (%s)", index, capacity));
         }
+    }
+
+    /**
+     * Creates a new VariableValueArray with the contents of this one, and the
+     * given bits per value.
+     * 
+     * @param newBitsPerValue
+     *            The new value. Must be larger than the current value (
+     *            {@link #getBitsPerValue()}).
+     * @throws IllegalArgumentException
+     *             If newBitsPerValue is less than or equal to the current bits
+     *             per value. Setting it to the same size would be a waste of
+     *             resources, and decreasing could lead to data loss.
+     * @return A new VariableValueArray
+     */
+    public VariableValueArray increaseBitsPerValueTo(int newBitsPerValue) {
+        if (newBitsPerValue < this.bitsPerValue) {
+            throw new IllegalArgumentException("Cannot decrease bits per value!  (was " + this.bitsPerValue + ", new size " + newBitsPerValue + ")");
+        } else if (newBitsPerValue == this.bitsPerValue) {
+            throw new IllegalArgumentException("Cannot resize to the same size!  (size was "  + newBitsPerValue + ")");
+        }
+
+        VariableValueArray returned = new VariableValueArray(newBitsPerValue, this.capacity);
+        for (int i = 0; i < this.capacity; i++) {
+            returned.set(i, this.get(i));
+        }
+        return returned;
+    }
+
+    @Override
+    public VariableValueArray clone() {
+        VariableValueArray clone = new VariableValueArray(this.bitsPerValue, this.capacity);
+        System.arraycopy(this.backing, 0, clone.backing, 0, this.backing.length);
+        return clone;
     }
 }
